Centralize registrations
