+++ apps/meteor/app/lib/lib/MessageTypes.ts
@@ -1,149 +1,4 @@
 import type { MessageTypesValues as MessageTypesValuesType } from '@rocket.chat/core-typings';
-import { MessageTypes } from '@rocket.chat/message-types';
-import { Meteor } from 'meteor/meteor';
-
-Meteor.startup(() => {
-	MessageTypes.registerType({
-		id: 'r',
-		system: true,
-		text: (t, message) => t('Room_name_changed_to', { room_name: message.msg }),
-	});
-	MessageTypes.registerType({
-		id: 'au',
-		system: true,
-		text: (t, message) => t('User_added_to', { user_added: message.msg }),
-	});
-	MessageTypes.registerType({
-		id: 'added-user-to-team',
-		system: true,
-		text: (t, message) => t('Added__username__to_this_team', { user_added: message.msg }),
-	});
-	MessageTypes.registerType({
-		id: 'ru',
-		system: true,
-		text: (t, message) => t('User_has_been_removed', { user_removed: message.msg }),
-	});
-	MessageTypes.registerType({
-		id: 'removed-user-from-team',
-		system: true,
-		text: (t, message) => t('Removed__username__from_the_team', { user_removed: message.msg }),
-	});
-	MessageTypes.registerType({
-		id: 'ult',
-		system: true,
-		text: (t) => t('User_left_this_team'),
-	});
-	MessageTypes.registerType({
-		id: 'user-converted-to-team',
-		system: true,
-		text: (t, message) => t('Converted__roomName__to_a_team', { roomName: message.msg }),
-	});
-	MessageTypes.registerType({
-		id: 'user-converted-to-channel',
-		system: true,
-		text: (t, message) => t('Converted__roomName__to_a_channel', { roomName: message.msg }),
-	});
-	MessageTypes.registerType({
-		id: 'user-removed-room-from-team',
-		system: true,
-		text: (t, message) => t('Removed__roomName__from_the_team', { roomName: message.msg }),
-	});
-	MessageTypes.registerType({
-		id: 'user-deleted-room-from-team',
-		system: true,
-		text: (t, message) => t('Deleted__roomName__room', { roomName: message.msg }),
-	});
-	MessageTypes.registerType({
-		id: 'user-added-room-to-team',
-		system: true,
-		text: (t, message) => t('Added__roomName__to_this_team', { roomName: message.msg }),
-	});
-	MessageTypes.registerType({
-		id: 'ujt',
-		system: true,
-		text: (t, message) => t('User_joined_the_team', { user: message.u.username }),
-	});
-	MessageTypes.registerType({
-		id: 'ut',
-		system: true,
-		text: (t, message) => t('User_joined_the_conversation', { user: message.u.username }),
-	});
-	MessageTypes.registerType({
-		id: 'wm',
-		system: true,
-		text: (t, message) => t('Welcome', { user: message.u.username }),
-	});
-	MessageTypes.registerType({
-		id: 'rm',
-		system: true,
-		text: (t) => t('Message_is_removed'),
-	});
-	MessageTypes.registerType({
-		id: 'user-muted',
-		system: true,
-		text: (t, message) => t('User_has_been_muted', { user_muted: message.msg }),
-	});
-	MessageTypes.registerType({
-		id: 'user-unmuted',
-		system: true,
-		text: (t, message) => t('User_has_been_unmuted', { user_unmuted: message.msg }),
-	});
-	MessageTypes.registerType({
-		id: 'subscription-role-added',
-		system: true,
-		text: (t, message) => t('set__username__as__role_', { username: message.msg, role: message.role ?? '' }),
-	});
-	MessageTypes.registerType({
-		id: 'subscription-role-removed',
-		system: true,
-		text: (t, message) => t('removed__username__as__role_', { username: message.msg, role: message.role ?? '' }),
-	});
-	MessageTypes.registerType({
-		id: 'room-archived',
-		system: true,
-		text: (t) => t('This_room_has_been_archived'),
-	});
-	MessageTypes.registerType({
-		id: 'room-unarchived',
-		system: true,
-		text: (t) => t('This_room_has_been_unarchived'),
-	});
-	MessageTypes.registerType({
-		id: 'room-removed-read-only',
-		system: true,
-		text: (t) => t('room_removed_read_only_permission'),
-	});
-	MessageTypes.registerType({
-		id: 'room-set-read-only',
-		system: true,
-		text: (t) => t('room_set_read_only_permission'),
-	});
-	MessageTypes.registerType({
-		id: 'room-allowed-reacting',
-		system: true,
-		text: (t) => t('room_allowed_reactions'),
-	});
-	MessageTypes.registerType({
-		id: 'room-disallowed-reacting',
-		system: true,
-		text: (t) => t('room_disallowed_reactions'),
-	});
-	MessageTypes.registerType({
-		id: 'room_e2e_enabled',
-		system: true,
-		text: (t) => t('Enabled_E2E_Encryption_for_this_room'),
-	});
-	MessageTypes.registerType({
-		id: 'room_e2e_disabled',
-		system: true,
-		text: (t) => t('Disabled_E2E_Encryption_for_this_room'),
-	});
-	MessageTypes.registerType({
-		id: 'videoconf',
-		system: false,
-		text: (t) => t('Video_Conference'),
-	});
-});
 
 export const MessageTypesValues: Array<{ key: MessageTypesValuesType; i18nLabel: string }> = [
 	{
