+++ packages/message-types/src/registrations/videoconf.ts
@@ -1,0 +1,9 @@
+import type { MessageTypes } from '../MessageTypes';
+
+export default (instance: MessageTypes) => {
+	instance.registerType({
+		id: 'videoconf',
+		system: false,
+		text: (t) => t('Video_Conference'),
+	});
+};
